feat(cli): store session token in OS keyring with file fallback
This change implements secure storage of the CLI token using the operating system keyring
with a fallback to plaintext file storage for Windows and macOS. Previously, the Coder
CLI stored authentication tokens in plaintext configuration files, which posed a security
risk because users' tokens are stored unencrypted and can be easily accessed by other
processes or users with file system access.

The secure storage is platform dependent. The security command is used on macOS.
Windows Credential Manager API is used on Windows. The session token continues to
be stored in plain text on Linux.