fix(dogfood): Pin docker-ce major version for containerd.io compat
This change pins the docker-ce major version for compatibility with the
containerd.io package version we have pinned. Newer versions of docker-ce
have a higher min version of containerd.io. While the underlying issue
that caused us to pin containerd.io is fixed in newer versions, we can't
yet remove the pin because some machines running dogfood are not yet updated
to a version of Linux that have a version of the fixed package (e.g.
ubuntu 20.04).