fix: retry embedded postgres port allocation
Sometimes tests would fail because the port embedded postgres tries to use
is already in use. This is because there's no way to tell postgres to use
an ephemeral port. This change adds retries to starting embedded postgres
when the port is not explicitly defined (e.g. tests), which should rid, or
at least significantly reduce, these flakes.