fix: ensure lifecycle executor has sufficient task permissions (#20539) (#20560)
We recently made a change to the `wsbuilder` to handle task related
logic. Our test coverage for the lifecycle executor didn't handle this
scenario and so we missed that it had insufficient permissions.

This PR adds `Update` and `Read` permissions for `Task`s in the
lifecycle executor, as well as an autostart/autostop test tailored to
task workspaces to verify the change.

---

This is cherry picked from
https://github.com/coder/coder/commit/06dbadab11760fe5fbf88c5bfcac2c48e11f7862
https://github.com/coder/coder/pull/20539