refactor!: remove TaskAppID from codersdk.WorkspaceBuild
We can instead use the new data model for this, there is no reason to
have it stored inside the workspace build. `TaskAppID` has not _yet_
been shipped into a release. This would need cherry-picking if we wanted
to go ahead with it.

If we don't want to cherry-pick, we should instead mark the TaskAppID as
deprecated. This isn't ideal as it hasn't even been released yet.