fix(site): fix task table links in Safari
This change makes the TableRow clickable rather than having an
absolutely positioned RouterLink. In Safari the entire table was broken
because all links spanned across the whole table, and the bottom most
row was moved to the top, resulting in only the bottom row being
highlighted and all rows leading to the bottom most task.