fix: send all params instead of only touched params (#20740)
resolves #20257

Ensure all parameters are sent in the web socket request instead of only
touched parameters.

Using touched parameters is irrelevant for the workspace parameters page
in workspace settings because parameters that appear here have already
been chosen by the user during workspace creation. So all parameters
should be sent in the web socket request whether have been touched in
the form or not.