fix: correctly set touched for autofill dynamic parameters (#20769)
this resolves #20187 

This bug was introduced in PR https://github.com/coder/coder/pull/18894
in CreateWorkspacePageViewExperimental.tsx

This PR made a change so that the entire parameter was used instead of
just the parameter name to mark an autofilled parameter as touched.

The fix here is to just use the parameter name.