fix(cli): don't default to keyring when --global-config set
This fixes a regression that caused the VS code extension to be unable
to authenticate after making keyring usage on by default. This is
because the VS code extension assumes the CLI will always use the session
token stored on disk, specifically in the directory specified by
--global-config.

This fix makes keyring usage enabled when the --global-config directory
is not set. This is a bit wonky but necessary to allow the extension to
continue working without modification and without backwards compat concerns.
In the future we should modify these extensions to either access the
credential in the keyring (like Coder Desktop) or some other approach that
doesn't rely on the session token being stored on disk.