feat(agent): add CLI commands to enable coder script interdependency (#20719)
relates to: https://github.com/coder/internal/issues/1094

This is number 5 of 5 pull requests in an effort to add agent script
ordering. It adds CLI commands that allow coder scripts to manage
dependencies between one another at runtime.

I used an LLM to produce some of these changes, but I have conducted
thorough self review and consider this contribution to be ready for an
external reviewer.

---------

Co-authored-by: Mathias Fredriksson <mafredri@gmail.com>