chore: add agentsocket api client (#20718)
relates to: https://github.com/coder/internal/issues/1094

This is number 4 of 5 pull requests in an effort to add agent script
ordering. It adds an API client that can call the workspace agent socket
API for script ordering operations.

In a follow-up PR, CLI commands will be added that consume this API.

I used an LLM to produce some of these changes, but I have conducted
thorough self review and consider this contribution to be ready for an
external reviewer.

---------

Co-authored-by: Mathias Fredriksson <mafredri@gmail.com>